==> 1. Data/00. Sources.R <==
# Cargar paquetes necesarios
library(tidyverse)
library(skimr)
library(openxlsx)

library(fixest) 
library(lmtest) 
library(dynlm) 
library(tseries)
library(broom)
library(spdep)
library(spatialreg)
library(splm)
library(ineptR)

library(sf)
library(geodata)
library(terra)

library(plm)
library(car)
library(systemfit)
library(modelsummary)
library(stargazer)
library(pandoc)

library(nlme)
library(GGally)

library(corrplot)
